backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
supabase==2.28.0
email-validator==2.1.2
openai>=1.0.0
